backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
celery
httpx
python-dateutil
pydantic
requests
lxml
python-multipart

# AI 和向量支持
openai>=1.0.0
pgvector
numpy

# 图片处理支持
Pillow>=9.0.0

# 文档解析支持
PyPDF2>=2.12.1,<4.0.0
Markdown>=3.4.1
chardet>=5.0.0
# python-magic>=0.4.27  # 可选，某些环境可能不需要

# 图片处理支持
Pillow>=9.0.0

# 可选：本地模型支持（CPU 环境可用）
# sentence-transformers
# torch
# transformers
